Added Decryption
This update adds functionality for decryption, instead of only being able to encrypt text. Users will now be expected to input three arguments. The first argument should be "enc" or "dec", to denote which mode they would like to run the program in. Then, they will input the mode and key, as in the previous version.